refactor: Moved packaging collection to be a repoClient specific sub-package
